Move driver to i2c::master
