+++ esp-hal/src/i2c/master/support/mod.rs
