+++ esp-hal/src/i2c/master/version/mod.rs
