+++ esp-hal/src/i2c/master/version/v1_v2_common.rs
@@ -1,4 +1,4 @@
-use crate::i2c::*;
+use crate::i2c::master::*;
 
 fn assert_read_length(len: usize) {
     if len <= 32 {
